Fixed warnings for 32 bit builds

+++ winpr/libwinpr/crt/test/TestAlignment.c
@@ -26,7 +26,7 @@ int TestAlignment(int argc, char* argv[])
 
 	if (((size_t) ptr % alignment) != 0)
 	{
-		printf("This pointer, %p, is not aligned on %lu\n", ptr, alignment);
+		printf("This pointer, %p, is not aligned on %lu\n", ptr, (unsigned long)alignment);
 		return -1;
 	}
 
@@ -36,7 +36,7 @@ int TestAlignment(int argc, char* argv[])
 	
 	if (((size_t) ptr % alignment) != 0)
 	{
-		printf("This pointer, %p, is not aligned on %lu\n", ptr, alignment);
+		printf("This pointer, %p, is not aligned on %lu\n", ptr, (unsigned long)alignment);
 		return -1;
 	}
 	
@@ -54,7 +54,8 @@ int TestAlignment(int argc, char* argv[])
 
 	if (((((size_t) ptr) + offset) % alignment) != 0)
 	{
-		printf("This pointer, %p, does not satisfy offset %lu and alignment %lu\n", ptr, offset, alignment);
+		printf("This pointer, %p, does not satisfy offset %lu and alignment %lu\n", ptr, (unsigned
+		long)offset, (unsigned long)alignment);
 		return -1;
 	}
 
@@ -70,7 +71,8 @@ int TestAlignment(int argc, char* argv[])
 
 	if (((((size_t) ptr) + offset) % alignment) != 0)
 	{
-		printf("This pointer, %p, does not satisfy offset %lu and alignment %lu\n", ptr, offset, alignment);
+		printf("This pointer, %p, does not satisfy offset %lu and alignment %lu\n", ptr, (unsigned
+		long)offset, (unsigned long)alignment);
 		return -1;
 	}
