+++ winpr/libwinpr/crt/test/TestString.c
@@ -38,7 +38,8 @@ int TestString(int argc, char* argv[])
 
 	if (length != testStringW_Length)
 	{
-		printf("_wcslen error: length mismatch: Actual: %lu, Expected: %lu\n", length, testStringW_Length);
+		printf("_wcslen error: length mismatch: Actual: %lu, Expected: %lu\n", (unsigned long) length,
+		(unsigned long) testStringW_Length);
 		return -1;
 	}
 
@@ -49,7 +50,8 @@ int TestString(int argc, char* argv[])
 
 	if (pos != 11)
 	{
-		printf("_wcschr error: position mismatch: Actual: %lu, Expected: %u\n", pos, 11);
+		printf("_wcschr error: position mismatch: Actual: %lu, Expected: %u\n", (unsigned long)pos,
+		11);
 		return -1;
 	}
 
@@ -58,7 +60,7 @@ int TestString(int argc, char* argv[])
 
 	if (pos != 29)
 	{
-		printf("_wcschr error: position mismatch: Actual: %lu, Expected: %u\n", pos, 29);
+		printf("_wcschr error: position mismatch: Actual: %lu, Expected: %u\n", (unsigned long)pos, 29);
 		return -1;
 	}
 
@@ -66,7 +68,8 @@ int TestString(int argc, char* argv[])
 
 	if (p != NULL)
 	{
-		printf("_wcschr error: return value mismatch: Actual: 0x%08lX, Expected: 0x%08lX\n", (size_t) p, (size_t) NULL);
+		printf("_wcschr error: return value mismatch: Actual: 0x%08lX, Expected: 0x%08lX\n", (unsigned
+		long) p, (unsigned long) NULL);
 		return -1;
 	}
